# specs/001-a-modular-high/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: ERP/Accounting Platform API
  description: |
    REST API for 1C-inspired ERP system with document-driven transactions,
    register-based accumulation, and automatic accounting entry generation.
  version: 1.0.0
  contact:
    name: API Support
    email: api@example.com
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://staging.example.com/api/v1
    description: Staging environment
  - url: https://api.example.com/api/v1
    description: Production environment
security:
  - BearerAuth: []
...
paths:
...
  /registers/inventory/movements:
    get:
      tags: [Registers]
      summary: Query inventory movements
      parameters:
        - name: itemId
          in: query
          schema:
            type: string
            format: uuid
        - name: warehouseId
          in: query
          schema:
            type: string
            format: uuid
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date-time
        - name: dateTo
          in: query
          schema:
            type: string
            format: date-time
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: List of inventory movements
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/InventoryMovement'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /registers/inventory/balances:
    get:
      tags: [Registers]
      summary: Query inventory balances
      parameters:
        - name: itemId
          in: query
          schema:
            type: string
            format: uuid
        - name: warehouseId
          in: query
          schema:
            type: string
            format: uuid
        - name: asOfDate
          in: query
          schema:
            type: string
            format: date-time
          description: Calculate balance as of this date (defaults to now)
      responses:
        '200':
          description: Inventory balances
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/InventoryBalance'

  # ========== REPORTS ==========
  /reports/inventory-balance/generate:
    post:
      tags: [Reports]
      summary: Generate inventory balance report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryBalanceReportConfig'
      responses:
        '200':
          description: Report generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryBalanceReport'
  /reports/inventory-balance/export/excel:
    post:
      tags: [Reports]
      summary: Export inventory balance report to Excel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryBalanceReportConfig'
      responses:
        '200':
          description: Excel file
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    MovementType:
      type: string
      enum: [RECEIPT, EXPENSE]
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer

    # Catalogs
...
    Item:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        description:
          type: string
        sku:
          type: string
        barcode:
          type: string
        unitOfMeasure:
          type: string
        isInventory:
          type: boolean
        isService:
          type: boolean
        defaultPrice:
          type: number
        costPrice:
          type: number
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Warehouse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        description:
          type: string
        address:
          type: string
        responsiblePerson:
          type: string
        isDefault:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    InventoryMovement:
      type: object
      properties:
        id:
          type: string
          format: uuid
        date:
          type: string
          format: date-time
        recorder:
          type: string
          format: uuid
        recordType:
          type: string
        item:
          $ref: '#/components/schemas/Item'
        warehouse:
          $ref: '#/components/schemas/Warehouse'
        quantity:
          type: number
        amount:
          type: number
        movementType:
          $ref: '#/components/schemas/MovementType'
    InventoryBalance:
      type: object
      properties:
        item:
          $ref: '#/components/schemas/Item'
        warehouse:
          $ref: '#/components/schemas/Warehouse'
        quantity:
          type: number
        amount:
          type: number

    # Reports
    InventoryBalanceReportConfig:
      type: object
      properties:
        asOfDate:
          type: string
          format: date-time
        warehouseId:
          type: string
          format: uuid
        itemId:
          type: string
          format: uuid
        minimumQuantity:
          type: number
    InventoryBalanceReport:
      type: object
      properties:
        config:
          $ref: '#/components/schemas/InventoryBalanceReportConfig'
        data:
          type: array
          items:
            $ref: '#/components/schemas/InventoryBalance'
        generatedAt:
          type: string
          format: date-time
...